+++ src/main/java/com/example/startlight/starList/repository/StarListRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface StarListRepository extends JpaRepository<StarList, Integer> {
+public interface StarListRepository extends JpaRepository<StarList, Long> {
 
     @Query("SELECT s from StarList s where s.pet.pet_id = :petId")
     List<StarList> findByPetId(@Param("petId") Long petId);
